Factor out Classes and methods to its own files
